Introduced System group that contains OS-level abstractions like URLSessionType
