+++ SwiftDemoApp/System/CollectionSafety.swift
